#176: renames and more ctors
