+++ src/main/java/org/cactoos/list/MappedIterator.java
@@ -23,34 +23,57 @@
  */
 package org.cactoos.list;
 
-import org.cactoos.ScalarHasValue;
-import org.hamcrest.MatcherAssert;
-import org.junit.Test;
+import java.util.Iterator;
+import org.cactoos.Func;
+import org.cactoos.func.UncheckedFunc;
 
 /**
- * Test case for {@link ConcatenatedIterable}.
+ * Mapped iterator.
+ *
+ * <p>There is no thread-safety guarantee.
+ *
  * @author Yegor Bugayenko (yegor256@gmail.com)
  * @version $Id$
+ * @param <X> Type of source item
+ * @param <Y> Type of target item
  * @since 0.1
- * @checkstyle JavadocMethodCheck (500 lines)
  */
-public final class ConcatenatedIterableTest {
+public final class MappedIterator<X, Y> implements Iterator<Y> {
 
-    @Test
-    @SuppressWarnings("unchecked")
-    public void transformsList() {
-        MatcherAssert.assertThat(
-            "Can't concatenate iterables together",
-            new LengthOfIterable(
-                new ConcatenatedIterable<>(
-                    new ArrayAsIterable<>("hello", "world", "друг"),
-                    new ArrayAsIterable<>("how", "are", "you"),
-                    new ArrayAsIterable<>("what's", "up")
-                )
-            ),
-            // @checkstyle MagicNumber (1 line)
-            new ScalarHasValue<>(8)
-        );
+    /**
+     * Iterator.
+     */
+    private final Iterator<X> iterator;
+
+    /**
+     * Function.
+     */
+    private final Func<X, Y> func;
+
+    /**
+     * Ctor.
+     * @param src Source iterable
+     * @param fnc Func
+     */
+    public MappedIterator(final Iterator<X> src, final Func<X, Y> fnc) {
+        this.iterator = src;
+        this.func = fnc;
+    }
+
+    @Override
+    public boolean hasNext() {
+        return this.iterator.hasNext();
+    }
+
+    @Override
+    @SuppressWarnings("PMD.AvoidCatchingGenericException")
+    public Y next() {
+        return new UncheckedFunc<>(this.func).apply(this.iterator.next());
+    }
+
+    @Override
+    public void remove() {
+        this.iterator.remove();
     }
 
 }
